requirements.txt
discord.py==0.16.11
aioodbc==0.2.0
python-decouple==3.1